add request and response DTOs and basic job fields validations

+++ internals/repositories/job_repository.go
@@ -12,7 +12,7 @@ func CreateJob(ctx context.Context, job models.Job) (models.Job, error) {
 	query := `
 INSERT INTO jobs(user_id,company_name,position,job_url,location,employment_type,salary_min,salary_max,description)
 VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9)
-RETURNING id,user_id,company_name,position,job_url,location,employment_type,salary_min,salary_max,description;
+RETURNING *;
 `
 	smt, err := db.DB.Prepare(query)
 
@@ -26,19 +26,19 @@ RETURNING id,user_id,company_name,position,job_url,location,employment_type,sala
 
 	var jobFromDB models.Job
 
-	err = row.Scan(&jobFromDB.Id, &jobFromDB.UserID, &jobFromDB.CompanyName, &jobFromDB.Position, &jobFromDB.JobURL, &jobFromDB.Location, &jobFromDB.EmploymentType, &jobFromDB.SalaryMin, &jobFromDB.SalaryMax, &jobFromDB.Description)
+	err = row.Scan(&jobFromDB.Id, &jobFromDB.UserID, &jobFromDB.CompanyName, &jobFromDB.Position, &jobFromDB.JobURL, &jobFromDB.Location, &jobFromDB.EmploymentType, &jobFromDB.SalaryMin, &jobFromDB.SalaryMax, &jobFromDB.Description, &jobFromDB.CreatedAT, &jobFromDB.UpdatedAT)
 
 	if err != nil {
-		return models.Job{}, customErr.ErrInternal
+		return models.Job{}, err
 	}
 
 	return jobFromDB, nil
 
 }
 
-func GetAllJobs(ctx context.Context, userId string) ([]models.Job, error) {
+func GetAllJobs(ctx context.Context, userId string) ([]models.JobResponse, error) {
 	query := `
-	SELECT * FROM jobs WHERE user_id=$1;
+	SELECT id,company_name,position,job_url,location,employment_type,salary_min,salary_max,description,created_at,updated_at FROM jobs WHERE user_id=$1;
 	`
 	rows, err := db.DB.QueryContext(ctx, query, userId)
 
@@ -48,43 +48,43 @@ func GetAllJobs(ctx context.Context, userId string) ([]models.Job, error) {
 
 	defer rows.Close()
 
-	var jobsList []models.Job
+	var jobsList []models.JobResponse
 	for rows.Next() {
-		var singleJob models.Job
+		var singleJob models.JobResponse
 
-		err := rows.Scan(&singleJob.Id, &singleJob.UserID, &singleJob.CompanyName, &singleJob.Position, &singleJob.JobURL, &singleJob.Location, &singleJob.EmploymentType, &singleJob.SalaryMin, &singleJob.SalaryMax, &singleJob.Description, &singleJob.CreatedAT, &singleJob.UpdatedAT)
+		err := rows.Scan(&singleJob.Id, &singleJob.CompanyName, &singleJob.Position, &singleJob.JobURL, &singleJob.Location, &singleJob.EmploymentType, &singleJob.SalaryMin, &singleJob.SalaryMax, &singleJob.Description, &singleJob.CreatedAT, &singleJob.UpdatedAT)
 
 		if err != nil {
-			return nil, customErr.ErrInternal
+			return nil, err
 		}
 
 		jobsList = append(jobsList, singleJob)
 	}
 
 	if err := rows.Err(); err != nil {
-		return nil, customErr.ErrInternal
+		return nil, err
 	}
 
 	return jobsList, nil
 }
 
-func GetJobByID(ctx context.Context, jobID string, userID string) (models.Job, error) {
+func GetJobByID(ctx context.Context, jobID string, userID string) (models.JobResponse, error) {
 	query := `
-SELECT * FROM jobs WHERE id=$1 AND user_id=$2;
+SELECT id,company_name,position,job_url,location,employment_type,salary_min,salary_max,description,created_at,updated_at FROM jobs WHERE id=$1 AND user_id=$2;
 `
 
 	row := db.DB.QueryRowContext(ctx, query, jobID, userID)
 
 	if err := row.Err(); err != nil {
-		return models.Job{}, err
+		return models.JobResponse{}, err
 	}
 
-	var singleJob models.Job
+	var singleJob models.JobResponse
 
-	err := row.Scan(&singleJob.Id, &singleJob.UserID, &singleJob.CompanyName, &singleJob.Position, &singleJob.JobURL, &singleJob.Location, &singleJob.EmploymentType, &singleJob.SalaryMin, &singleJob.SalaryMax, &singleJob.Description, &singleJob.CreatedAT, &singleJob.UpdatedAT)
+	err := row.Scan(&singleJob.Id, &singleJob.CompanyName, &singleJob.Position, &singleJob.JobURL, &singleJob.Location, &singleJob.EmploymentType, &singleJob.SalaryMin, &singleJob.SalaryMax, &singleJob.Description, &singleJob.CreatedAT, &singleJob.UpdatedAT)
 
 	if err != nil {
-		return models.Job{}, customErr.ErrJobNotFound
+		return models.JobResponse{}, customErr.ErrJobNotFound
 	}
 
 	return singleJob, nil
@@ -94,8 +94,8 @@ func UpdateJob(ctx context.Context, jobID string, job models.Job) (models.Job, e
 
 	query := `
 	UPDATE jobs
-	SET company_name=$2,position=$3,job_url=$4,location=$5,employment_type=$6,salary_min=$7,salary_max=$8,description=$9,updated_at=CURRENT_TIMESTAMP
-	WHERE id=$1
+	SET company_name=$3,position=$4,job_url=$5,location=$6,employment_type=$7,salary_min=$8,salary_max=$9,description=$10,updated_at=CURRENT_TIMESTAMP
+	WHERE id=$1 AND user_id=$2
 	RETURNING *;
 	`
 
@@ -106,34 +106,34 @@ func UpdateJob(ctx context.Context, jobID string, job models.Job) (models.Job, e
 
 	defer smt.Close()
 
-	row := smt.QueryRowContext(ctx, jobID, job.CompanyName, job.Position, job.JobURL, job.Location, job.EmploymentType, job.SalaryMin, job.SalaryMax, job.Description)
+	row := smt.QueryRowContext(ctx, jobID, job.UserID, job.CompanyName, job.Position, job.JobURL, job.Location, job.EmploymentType, job.SalaryMin, job.SalaryMax, job.Description)
 
 	if err := row.Err(); err != nil {
 		return models.Job{}, err
 	}
 
-	var singleJob models.Job
+	var jobFromDB models.Job
 
-	err = row.Scan(&singleJob.Id, &singleJob.UserID, &singleJob.CompanyName, &singleJob.Position, &singleJob.JobURL, &singleJob.Location, &singleJob.EmploymentType, &singleJob.SalaryMin, &singleJob.SalaryMax, &singleJob.Description, &singleJob.CreatedAT, &singleJob.UpdatedAT)
+	err = row.Scan(&jobFromDB.Id, &jobFromDB.UserID, &jobFromDB.CompanyName, &jobFromDB.Position, &jobFromDB.JobURL, &jobFromDB.Location, &jobFromDB.EmploymentType, &jobFromDB.SalaryMin, &jobFromDB.SalaryMax, &jobFromDB.Description, &jobFromDB.CreatedAT, &jobFromDB.UpdatedAT)
 
 	if err != nil {
 		return models.Job{}, err
 	}
 
-	return singleJob, nil
+	return jobFromDB, nil
 
 }
 
-func DeleteJob(ctx context.Context, JobID string) error {
+func DeleteJob(ctx context.Context, JobID, userId string) error {
 	query := `
-	DELETE FROM jobs WHERE id=$1;
+	DELETE FROM jobs WHERE id=$1 AND user_id=$2;
 	`
 
-	_, err := db.DB.ExecContext(ctx, query, JobID)
+	_, err := db.DB.ExecContext(ctx, query, JobID, userId)
 
 	if err != nil {
 
-		return customErr.ErrInternal
+		return err
 	}
 	return nil
